+++ src/index.ts
@@ -39,6 +39,6 @@ export {
   NetworkType
 } from "./negotiation/execution_params";
 export { MakerNegotiator, MakerHttpApi } from "./negotiation/maker_negotiator";
-export { TakerNegotiator, MakerClient } from "./negotiation/taker_negotiator";
+export { TakerNegotiator } from "./negotiation/taker_negotiator";
 
 export { TryParams } from "./timeout_promise";